Remove Null Atom
