+++ radixdlt-java/src/main/java/com/radixdlt/client/core/serialization/SerializedAtomType.java
@@ -1,13 +1,11 @@
 package com.radixdlt.client.core.serialization;
 
 import com.radixdlt.client.core.atoms.Atom;
-import com.radixdlt.client.core.atoms.NullAtom;
 import com.radixdlt.client.core.atoms.TransactionAtom;
 import java.util.Optional;
 
 public enum SerializedAtomType {
-	TRANSACTION(TransactionAtom.class, -760130L),
-	NULL(NullAtom.class, -1123323048L);
+	TRANSACTION(TransactionAtom.class, -760130L);
 
 	private final Class<? extends Atom> atomClass;
 	private final long serializer;
